requirements.txt
# local package
-e .

# external requirements
python
numpy
pandas
matplotlib
tensorflow
tensorflow_hub
PIL.Image
